requirements.txt
streamlit==1.28.1
reportlab==4.0.5
Pillow==10.0.1
pandas==2.1.3
numpy==1.24.3
pytesseract==0.3.10
pdf2image==1.16.3
openai==1.3.0
python-dotenv==1.0.0
openpyxl==3.1.2
pymupdf==1.23.8
easyocr==1.7.0
xlsxwriter==3.1.9 